refactor code based on last changes to common submodule
