+++ package.json
@@ -10,11 +10,11 @@
     "build": "quasar build",
     "cypress:open:dev": "cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress open",
     "cypress:open": "cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress open",
-    "cypress:integration:ci": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=integration",
-    "cypress:integration:ci:dev": "yarn cypress:integration:ci",
+    "cypress:integration": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=integration",
+    "cypress:integration:dev": "yarn cypress:integration",
     "cypress:e2e": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=e2e",
-    "start:integration:ci": "concurrently \"yarn cypress:integration:ci\" --kill-others \"yarn start\" --success first",
-    "start:integration:ci:dev": "concurrently \"yarn cypress:integration:ci:dev\" --kill-others \"yarn start:dev\" --success first",
+    "start:integration": "concurrently \"yarn cypress:integration\" --kill-others \"yarn start\" --success first",
+    "start:integration:dev": "concurrently \"yarn cypress:integration:dev\" --kill-others \"yarn start:dev\" --success first",
     "start:e2e": "concurrently \"yarn cypress:e2e\" --kill-others \"yarn start\" --success first",
     "lint": "eslint --ext .js,.vue src",
     "start": "quasar serve dist/spa -H '0.0.0.0' -p 4000",
@@ -81,7 +81,6 @@
     "eslint-plugin-vue": "^5.2.2",
     "husky": "^2.3.0",
     "jest-junit": "^6.4.0",
-    "json-graphql-server": "^2.1.3",
     "jwt-decode": "^2.2.0",
     "majestic": "^1.4.1",
     "moment": "^2.24.0",
